除分页外的所有基本功能
类型管理中 删除类型 只能删除未被商品占用的类型
商品管理中 搜索框目前只能根据类型管理中的类型ID查找商品

+++ frontend/src/pages/Goods/Allproducts.js
@@ -1,4 +1,4 @@
-import React,{useEffect, useState,useRef} from 'react';
+import React, { useEffect, useState, useRef } from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import InputBase from '@material-ui/core/InputBase';
 import Button from '@material-ui/core/Button';
@@ -12,7 +12,11 @@ import DetailsPage from "./DetailsPage";
 import config from "../../config/config.json";
 import axios from "axios";
 import Home from "../../View/Home";
-import {Route} from "react-router-dom";
+import { Route } from "react-router-dom";
+import Snackbar from "../../components/Snackbar/Snackbar";
+import {AddAlert} from "_@material-ui_icons@4.11.2@@material-ui/icons";
+import GridItem from "../../components/Grid/GridItem";
+import Pagination from '@material-ui/lab/Pagination';
 
 
 const useStyles = makeStyles((theme) => ({
@@ -39,17 +43,63 @@ export default function Inputs(props) {
     const [dialogTitle, setDialogTitle] = useState("查看公私钥")     //设置对话框提示内容
     const [dialogTitleq, setDialogTitleq] = useState("查看公私钥")     //设置对话框提示内容
     const [abs, setAbs] = useState([]);
-    const [proId,setProId] =useState({});
-    const [pro,setPro] =useState({});
-    const [det,setDet] =useState({});
+    const [proId, setProId] = useState({});
+    const [pro, setPro] = useState({});
+    const [det, setDet] = useState({});
 
-    useEffect(()=>{
-        headen()
-    },[])
-    const change=(item,id)=>{
+    const [tc, setTC] = React.useState(false);             //设置提示框的显示隐藏
+    const [message, setMessage] = React.useState("");            //设置提示框的提示信息
+
+    const [pag, setPag] = useState([]);
+    const [page, setPage] = React.useState(1);
+
+    useEffect(() => {
+        headen();
+        // PAGE(page)
+    }, [])
+
+    // const handlePageChange = (event,value) => {
+    //     setPage(value)
+    //     PAGE(value)
+    // }
+
+    // const PAGE = (a) => {
+    //     axios.get('http://192.168.0.124:8001/commodity', {
+    //         params: {
+    //             page: a,
+    //             limit: 5,
+    //         }
+    //     })
+    //         .then((res) => {
+    //             console.log(res.data)
+    //             setAbs(
+    //                 res.data.data,
+    //             );
+    //             setPag(
+    //                 res.data.page_count
+    //             )
+    //         })
+    //         .catch((err) => {
+    //             console.log('err')
+    //         })
+    // }
+
+    const flagSnackbar = (messages) => {
+        if (!tc) {
+            setTC(true)
+            setMessage(messages)
+            setTimeout(function () {
+                setTC(false);
+                setMessage("")
+            }, 6000);
+        }
+    }
+
+
+    const change = (item, id) => {
         return (
             <div>
-                <Button variant="outlined" color="primary" onClick={()=>{
+                <Button variant="outlined" color="primary" onClick={() => {
                     setDialogTitle("修改商品信息");
                     setOpen(true);
                     setTitle(true);
@@ -61,29 +111,34 @@ export default function Inputs(props) {
                 <Button
                     variant="outlined"
                     color="secondary"
-                    onClick={()=>{Delete(id)}}
-                    style={{marginLeft:'10px'}}>
+                    onClick={() => { Delete(id) }}
+                    style={{ marginLeft: '10px' }}>
                     删除
                 </Button>
             </div>
         )
     }
-    const Delete=(id)=>{
-        axios.delete(config.httpUrlpro1+id)
-            .then((res) => {
-                console.log(res.data);
-                window.history.go(0)
+    const Delete = (id) => {
+        // eslint-disable-next-line no-restricted-globals
+        if(confirm('是否确认删除？')) {
+            axios.delete(config.httpUrlpro1 + id)
+                .then((res) => {
+                    console.log(res.data);
+                    flagSnackbar('删除成功')
+                    window.history.go(0)
 
-            })
-            .catch((err) => {
-                console.log('err')
-            })
+                })
+                .catch((err) => {
+                    console.log('err')
+                })
+        }
+
     }
 
     const inputData = useRef(null);
 
     const headen = () => {
-        axios.get( config.httpUrlpro1 )
+        axios.get(config.httpUrlpro1)
             .then((res) => {
                 setAbs(
                     res.data
@@ -94,10 +149,10 @@ export default function Inputs(props) {
             })
     }
 
-    const details =(id)=> {
-        return(
+    const details = (id) => {
+        return (
             <div>
-                <Button variant="outlined" color="inherit" onClick={()=>{
+                <Button variant="outlined" color="inherit" onClick={() => {
                     // deta(id);
                     setDet(id)
                     setDialogTitleq("修改商品信息");
@@ -116,14 +171,14 @@ export default function Inputs(props) {
             for (let i = 0; i < tableJson.length; i++) {
                 tabData.push(
                     [
-                        i+1,
+                        i + 1,
                         // tableJson[i].commodity_id,
                         tableJson[i].commodity_name,
                         tableJson[i].quantity_in_stock,
                         // tableJson[i].type,
                         tableJson[i].remark,
                         // details(tableJson[i].commodity_id),
-                        change(tableJson[i],tableJson[i].commodity_id),
+                        change(tableJson[i], tableJson[i].commodity_id),
                         details(tableJson[i].commodity_id)
                         // <SearchTwoToneIcon color="secondary" onClick={()=>{setMerchantOneItem(tableJson[i]);setOpen(true);setDialogTitle("查看公私钥")}} className={classes.pointer} titleAccess="查看公钥"/>
                     ]
@@ -139,42 +194,32 @@ export default function Inputs(props) {
         }
         return tabData
     }
-    function dialogOpen(){
+    function dialogOpen() {
         setOpen(false);
         setOpenq(false);
         window.history.go(0);
     }
-    function dialogOpenq(){
+    function dialogOpenq() {
         setOpenq(false);
     }
 
-
-    // const blurFind = ()=>{
-    //     let keyword = prompt("请输入查询关键字！")
-    //     let result = tableData.filter((item,i)=>{
-    //         if(item.includes(keyword)){
-    //             return true
-    //         }
-    //         return false
-    //     })
-    //     //实现关键词高亮
-    //     result.forEach((item,i,arr)=>{
-    //         arr[i] = arr[i].replace(new RegExp(keyword,"g"),"<span style='color:red'>"+keyword+"</span>")
-    //     })
-    //     this.setState({
-    //         arr:result,
-    //         flag:false  //更改状态
-    //     })
-    // }
-
-
-
     return (
-        <div style={{position:'relative',margin:'0px 50px'}}>
+        <div style={{ position: 'relative', margin: '0px 50px' }}>
+            <GridItem xs={12} sm={12} md={4}>
+                <Snackbar
+                    place="tc"
+                    color="info"
+                    icon={AddAlert}
+                    message={message}
+                    open={tc}
+                    closeNotification={() => setTC(false)}
+                    close
+                />
+            </GridItem>
             {/*删除修改页*/}
             <Dialog
                 open={open}
-                onClose={()=>{dialogOpen()}}
+                onClose={() => { dialogOpen() }}
                 aria-labelledby="scroll-dialog-title"
                 aria-describedby="scroll-dialog-description"
             >
@@ -203,7 +248,7 @@ export default function Inputs(props) {
                     }
                 </DialogContent>
                 <DialogActions>
-                    <Button onClick={()=>dialogOpen()} color="primary" variant="outlined">
+                    <Button onClick={() => dialogOpen()} color="primary" variant="outlined">
                         关闭
                     </Button>
                 </DialogActions>
@@ -211,7 +256,7 @@ export default function Inputs(props) {
             {/*详情页*/}
             <Dialog
                 open={openq}
-                onClose={()=>{dialogOpen()}}
+                onClose={() => { dialogOpen() }}
                 aria-labelledby="scroll-dialog-title"
                 aria-describedby="scroll-dialog-description"
             >
@@ -239,61 +284,68 @@ export default function Inputs(props) {
                     }
                 </DialogContent>
                 <DialogActions>
-                    <Button onClick={()=>dialogOpenq()} color="primary" variant="outlined">
+                    <Button onClick={() => dialogOpenq()} color="primary" variant="outlined">
                         关闭
                     </Button>
                 </DialogActions>
             </Dialog>
-            <div style={{boxSizing:'border-box',marginBottom:'50px'}}>
+            <div style={{ boxSizing: 'border-box', marginBottom: '50px' }}>
                 {/*输入框*/}
                 <div className={classes.root}
-                      style={{
-                          width:'100%',
-                          padding:'2px 4px',
-                          display:'flex',
-                          backgroundColor:'#ab47bc',
-                      }}
-                      noValidate autoComplete="off"
+                    style={{
+                        width: '100%',
+                        padding: '2px 4px',
+                        display: 'flex',
+                        backgroundColor: '#ab47bc',
+                    }}
+                    noValidate autoComplete="off"
                 >
                     <InputBase
                         className={classes.margin}
                         placeholder="请输入类型管理中的类型ID查询"
                         style={{
                             // flex:'1',
-                            backgroundColor:'white',
-                            padding:'6px 0 7px 6px',
-                            border:'1px solid black'
+                            backgroundColor: 'white',
+                            padding: '6px 0 7px 6px',
+                            border: '1px solid black'
                         }}
                         fullWidth={true}
                         inputProps={{ 'aria-label': 'naked' }}
-                        onKeyDown={(data)=>{
-                            if(data.key==='Enter'){
+                        onKeyDown={(data) => {
+                            if (data.key === 'Enter') {
                                 console.log(inputData.current.lastChild.value);
                                 const typeid = inputData.current.lastChild.value;
-                                axios.get(config.httpUrlType+'?type_id='+typeid,{headers:{}})
-                                    .then((res)=>{
+                                axios.get(config.httpUrlType + '?type_id=' + typeid, { headers: {} })
+                                    .then((res) => {
                                         console.log(res)
                                         setAbs(res.data)
                                     })
-                                    .catch((err)=>{
+                                    .catch((err) => {
                                         console.log('err')
                                     })
-                            }}}
+                            }
+                        }}
                         ref={inputData}
                     />
                 </div>
             </div>
             {/*添加按钮*/}
-            <div style={{position:'absolute',top:'90px', right:'30px'}}>
+            <div style={{ position: 'absolute', top: '90px', right: '30px' }}>
                 <div className={classes.rightStyle}>
-                    <Button variant="outlined" color="primary" onClick={()=>{setDialogTitle("添加商品信息");setOpen(true);setTitle(false);}}>添加</Button>
+                    <Button variant="outlined" color="primary" onClick={() => { setDialogTitle("添加商品信息"); setOpen(true); setTitle(false); }}>添加</Button>
                 </div>
             </div>
             {/*列表渲染*/}
             <Table
-                tableHead={['序号','商品名称','库存数量','备注','操作','详情']}
+                tableHead={['序号', '商品名称', '库存数量', '备注', '操作', '详情']}
                 tableData={tableData()}
             />
+            <div className={classes.root}>
+                <Pagination count={5} variant="outlined" color="primary"
+                            // onChange={handlePageChange}
+                />
+                共<span>{5}</span>页
+            </div>
         </div>
     );
 }
